requirements.txt
langchain>=0.3.0
langgraph>=0.3.0
langchain-google-genai>=4.0.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
langsmith>=0.2.0
pydantic>=2.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
